fix: move chatDb file, add docstrings
